+++ sample_app/app/jobs/hello_job.rb
@@ -1,7 +1,18 @@
 class HelloJob < ApplicationJob
   queue_as :default
 
+  class NameException < StandardError; end
+
+  class SkipException < StandardError; end
+
+  retry_on NameException
+  discard_on SkipException
+
   def perform(name)
+    raise NameException if name == "error"
+    raise SkipException if name == "skip"
+    raise StandardError if name == "StandardError"
+
     puts "Hello from our job: #{name}"
   end
 end